requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.116.0
filelock==3.18.0
Flask==3.1.1
Flask-RESTful==0.3.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
magic-filter==1.0.12
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.6.3
networkx==3.2.1
numpy==1.26.4
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
psutil==7.0.0
py-cpuinfo==9.0.0
pybind11==2.12.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scipy
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.163
ultralytics-thop==2.0.14
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
